transfer project to twitter clone
i've decided to transform the projectto a twitter clone social app instead of my own UI/UX in order to build the app faster. my main goal i want to achieve out of this project is to improve my developing skills with less focus on design and user experience. i've altered the project docx as well as dome data types on the models