converted all the stateless widgets that uses the api.dart functions to statefull
we dont want our widgets to call the api every time they get rebuild. so in order to fix that we can convert each of these widget to a statefull one and instead of calling the api from the build method, we will call it within initState(). that will assure that the api call will happen only once, when the widget is build at the first time.